move domain block er protocol to sc-domains and register the protocol to consensus node
