+++ crates/sc-domains/Cargo.toml
@@ -16,9 +16,16 @@ include = [
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+array-bytes.workspace = true
+async-channel.workspace = true
+domain-runtime-primitives.workspace = true
 frame-benchmarking = { workspace = true, optional = true }
+futures.workspace = true
+parity-scale-codec.workspace = true
 sc-client-api.workspace = true
 sc-executor.workspace = true
+sc-network.workspace = true
+sc-network-sync.workspace = true
 sp-api.workspace = true
 sp-auto-id = { workspace = true, features = ["std"] }
 sp-blockchain.workspace = true
@@ -30,6 +37,9 @@ sp-io.workspace = true
 sp-messenger-host-functions = { workspace = true, features = ["std"] }
 sp-runtime.workspace = true
 sp-subspace-mmr.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tracing.workspace = true
 
 [features]
 runtime-benchmarks = ["dep:frame-benchmarking"]
