+++ crates/sc-domains/src/domain_block_er.rs
@@ -1,0 +1,4 @@
+//! Domain Block ER request response protocol.
+
+pub mod execution_receipt_protocol;
+pub mod receipt_receiver;
