+++ domains/service/Cargo.toml
@@ -13,7 +13,6 @@ build = "build.rs"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-array-bytes.workspace = true
 async-channel.workspace = true
 async-trait.workspace = true
 cross-domain-message-gossip.workspace = true
