+++ domains/service/src/network.rs
@@ -1,4 +1,3 @@
-use crate::network::execution_receipt_protocol::DomainBlockERRequestHandler;
 use sc_client_api::{AuxStore, BlockBackend, BlockchainEvents, ProofProvider};
 use sc_consensus::ImportQueue;
 use sc_network::NetworkBackend;
@@ -17,19 +16,14 @@ use sp_api::ProvideRuntimeApi;
 use sp_api::__private::BlockT;
 use sp_blockchain::{HeaderBackend, HeaderMetadata};
 use sp_consensus::block_validation::{Chain, DefaultBlockAnnounceValidator};
-use sp_domains::DomainsApi;
-use sp_runtime::traits::{BlockIdTo, Header};
+use sp_runtime::traits::BlockIdTo;
 use std::sync::Arc;
-
-pub mod execution_receipt_protocol;
-pub mod receipt_receiver;
 
 /// Build the network service, the network status sinks and an RPC sender.
 #[allow(clippy::type_complexity)]
 #[expect(clippy::result_large_err, reason = "Comes from Substrate")]
-pub fn build_network<Block, Net, TxPool, IQ, Client, CBlock, CClient, DomainHeader>(
+pub fn build_network<Block, Net, TxPool, IQ, Client>(
     params: BuildNetworkParams<Block, Net, TxPool, IQ, Client>,
-    consensus_client: Arc<CClient>,
 ) -> Result<
     (
         Arc<dyn sc_network::service::traits::NetworkService>,
@@ -43,8 +37,7 @@ pub fn build_network<Block, Net, TxPool, IQ, Client, CBlock, CClient, DomainHead
     Error,
 >
 where
-    Block: BlockT<Header = DomainHeader>,
-    CBlock: BlockT,
+    Block: BlockT,
     Client: ProvideRuntimeApi<Block>
         + HeaderMetadata<Block, Error = sp_blockchain::Error>
         + Chain<Block>
@@ -58,15 +51,6 @@ where
     TxPool: TransactionPool<Block = Block, Hash = <Block as BlockT>::Hash> + 'static,
     IQ: ImportQueue<Block> + 'static,
     Net: NetworkBackend<Block, <Block as BlockT>::Hash>,
-    CClient: ProvideRuntimeApi<CBlock>
-        + BlockBackend<CBlock>
-        + ProofProvider<CBlock>
-        + HeaderBackend<CBlock>
-        + Send
-        + Sync
-        + 'static,
-    CClient::Api: DomainsApi<CBlock, DomainHeader>,
-    DomainHeader: Header,
 {
     let BuildNetworkParams {
         config,
@@ -87,26 +71,6 @@ where
     } else {
         Box::new(DefaultBlockAnnounceValidator)
     };
-
-    // "Last confirmed domain block execution receipt" request handler
-    {
-        let num_peer_hint = net_config.network_config.default_peers_set_num_full as usize
-            + net_config
-                .network_config
-                .default_peers_set
-                .reserved_nodes
-                .len();
-
-        let (handler, protocol_config) =
-            DomainBlockERRequestHandler::new::<Net>(fork_id, consensus_client, num_peer_hint);
-        spawn_handle.spawn(
-            "last-domain-execution-receipt-request-handler",
-            Some("networking"),
-            handler.run(),
-        );
-
-        net_config.add_request_response_protocol(protocol_config);
-    }
 
     let network_service_provider = NetworkServiceProvider::new();
     let protocol_id = config.protocol_id();
